+++ source/test-suite/fixtures/TemplateHaskellFallbackExpected.hs
@@ -1,0 +1,5 @@
+{-# LANGUAGE TemplateHaskell #-}
+module TemplateHaskellFallback where
+
+quoteExpVersion txt = [|parseVersionUnsafe txt|]
+  where parseVersionUnsafe value = value