+++ source/test-suite/fixtures/TemplateHaskellFallbackInvalid.hs
@@ -1,0 +1,4 @@
+{-# LANGUAGE TemplateHaskell #-}
+module TemplateHaskellFallbackInvalid where
+
+broken = [|unterminated